requirements.txt
discord.py
requests
beautifulsoup4
os
asyncio
mcstatus
random
datetime
sqlite3
aiohttp